# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Customer API
  description: API for managing customers
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
...
  /customer/contact:
    post:
      summary: Add contact to customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContactOrDocumentDTO'
      responses:
        '200':
          description: Contact added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerContactDTO'
  /customer/document:
    post:
      summary: Add document to customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddContactOrDocumentDTO'
      responses:
        '200':
          description: Document added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDocumentDTO'
components:
  schemas:
...
    CustomerContactDTO:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
    CustomerDocumentDTO:
      type: object
      properties:
        documentNumber:
          type: string
        documentType:
          type: string
    AddContactOrDocumentDTO:
      type: object
      properties:
        type:
          type: string
        value:
          type: string
        customerId:
          type: integer